Remove InProtectedSequence check from Layer::IsPropertyChangeAllowed
This was originally added here, before there were any actual thread
protections:

https://chromium-review.googlesource.com/c/chromium/src/+/3207371

Now that thread protections for commit are in place, this is no
longer the desired behavior. It's OK to modify a property on the
main thread while commit is still running on the impl thread,
because the Layer code will block as necessary to wait for commit to
finish.

Bug: chromium:1255972
Change-Id: I9cd1d0e8a8f58862d0b6b252036d5cabbc26df33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3468980
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972075}