ash: Distinguish between individual calendar event fetches
Each individual fetch of a month's calendar events now reports its own
status, and a given month is officially marked as "fetched" only if
its fetch was successful.  At the moment we don't explicitly retry a
failed event fetch, but if it wasn't marked as "fetched" then we'll
try it again next time the user browses to it.

Bug: 1258179
Change-Id: I249ddac5a72026c36ec870fadbe7b58d60998ccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3448744
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Roger Tinkoff <rtinkoff@google.com>
Cr-Commit-Position: refs/heads/main@{#972079}