[CrOS Bluetooth] Fix GetPairedDeviceName encoding conversion.
Update CrosBluetoothConfigUtil::GetPairedDeviceName() to convert
nicknames from UTF-8 to UTF-16. Previously, it treated nicknames as
ASCII when converting to UTF-16. However, when emojis were used, this
caused a DCHECK() to fail as emojis cannot be represented in ASCII.
Because mojo strings are UTF-8, this method was updated to treat
nicknames as UTF-8 when converting to UTF-16.

Screenshots:
https://screenshot.googleplex.com/9yRNHDSz6v6k6Y5.png
https://screenshot.googleplex.com/AE9uSNUTUro3zaU.png
https://screenshot.googleplex.com/zZqawkaQfvB6dTc.png

Fixed: b/219525413
Change-Id: Id32c108d8907c1ab1a622fff979490f7f441f37f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469659
Commit-Queue: Gordon Seto <gordonseto@google.com>
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972104}