Make RequestSocket() non-reentrant
Currently, TransportClientSocketPool::RequestSocket() calls itself
re-entrantly when Secure DNS is enabled, in order to acquire a socket to
the DoH server.  RequestSocket ultimately interacts with
ResolveHostRequest, which is not reentrancy-safe, leading to a DCHECK
when Secure DNS Enhanced Bootstrap is enabled.

This change:
1. Makes Secure DNS socket acquisition asynchronous to break the
   re-entrance cycle.
2. Removes logic from RequestSocket() for tolerating re-entrant usages.
3. Adds a DCHECK to ensure that RequestSocket() has no re-entrant calls.

This partly reverse changes originally made in
https://chromium-review.googlesource.com/c/chromium/src/+/1506567

BUG=1200908

Change-Id: I57df0530df033dc2ca399f05808bbab48fa63172
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3457644
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Commit-Queue: Ben Schwartz <bemasc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972121}