[Extensions] Fix executeScript on new tab with uncommitted navigation
This CL addresses a problem that could occur if scripting.executeScript
was called on a newly created tab on which the navigation had not
committed on yet. It adds in a case when checking if the extension has
permission to the frame, to leave it up to the renderer to check again
later if the last committed URL on the frame is not valid. This check
existed in the older executeScript APIs, but was missing from the new
scripting API version.

Also adds an associated test.

Bug: 1191971
Change-Id: Idc0ca0cd33560dcc53fa248050f277b6b6eb93fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3413666
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972137}