[Launcher Sort] Handle item icon update in color order
This CL ensures that under color sorting when an item's icon updates,
the item's position changes accordingly based on the color order. In
detail, this CL does the following things:

(1) Refactor the code so that in AppListItem, AppListModel and
ChromeAppListModelUpdater when setting an item's icon, set the item's
icon color at the same time. In contrast, currently icon and icon color
are set with different methods.

(2) In ChromeAppListModelUpdater::SetItemIconAndColor(), check whether
the app list is under color sorting (including the temporary order
and the permanent order). If so, update the item's position if
necessary.

(3) Update IconColorWrapperComparator() so that color sorting becomes
stable.

(4) Create browser test cases to verify two scenarios:
    (i) Change the icon of an item already added to the model updater
when the app list is sorted by color.
    (ii) Add a new item to the model updater when the app list is sorted
by color.

Bug: 1293162
Change-Id: Iffb14a7543e32b01d8fc237ccfa4d45f1f876986
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3459802
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972166}