personalization: take webcam photo as avatar
Allow user to take a webcam photo from personalization app and save as
device avatar image. Capture the image in JS and encode as PNG.
Transfer over mojom as a BigBuffer of the png data, and set as user
image if the bytes are successfully decoded.

BUG=b:202877554
TEST=browser_tests --gtest_filter="PersonalizationAppComponent*"
TEST=open personalization app, change avatar, open camera, snap photo

Change-Id: I378579c9b03394b740b9730956d1f40f37084720
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3438566
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Jeffrey Young <cowmoo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972168}