Refactor VirtualCardEnrollmentManager and PaymentsClient for the Clank settings page use case, and remove error dialog on GetDetailsForEnroll response
This CL sets up the new structure for VirtualCardEnrollmentManager and PaymentsClient so that it can handle the Clank settings page use-case as well.

It also removes the showing of the error dialog for a failed GetDetailsForEnroll response, so that the user does not have a confusing user experience if they see an error dialog for a flow they did not initiate.

Bug: 1281695
Change-Id: I6f4864af3a8847cbd639b609029e289f9bfc075c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3448202
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: Jared Saul <jsaul@google.com>
Reviewed-by: Siddharth Shah <siashah@chromium.org>
Reviewed-by: Siyu An <siyua@chromium.org>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Cr-Commit-Position: refs/heads/main@{#972181}