Add a way for embedders to ignore initial NavigationEntry NavigationStateChanged calls
WebView should not fire onPageFinished calls in response to initital
NavigationEntry creation/update, as those entries used to not exist
before. We added a way for WebView to do that in a hacky way through
passing the information with the InvalidateType in crrev.com/c/3347483.
This CL removes the hack and adds a way for the embedders to tell
the content/ code to not fire the NavigationStateChanged calls at
all, by adding a method for that in ContentBrowserClient.

This CL is not changing any behavior, just moves the decision-making
to content/ code. Note that we might remove the method entirely in the
future if the approach at crrev.com/c/3428751 works, which removes the
need to add this method without having to send additional information
from content/ to WebView code.

See also:
https://groups.google.com/a/chromium.org/g/content-owners/c/D39aztdWLxY/m/Zk4cSuf6AAAJ

Bug: 1293332
Change-Id: Ib214cede6fa68627d430336930e072009e26db9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3427984
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Charles Reis <creis@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972187}