[journeys] Introduce `url_for_deduping` which preserves path and query.
This means that many fewer visits will be marked as duplicates by
Journeys, when using the SimilarVisitDeduperClusterFinalizer.

But more visits will be marked as duplicates by the
UrlDeduperClusterFinalizer, because it previously used the full
`normalized_url`.

It makes the two more similar by pushing the URL-canonicalization into
a utility function similar to how Omnibox does it.

Bug: 1292134, 1294350
Change-Id: I050206c072046c529245ee0b4609445f95ca986c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3468989
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972192}