[CrOS Bluetooth] Fix flaky BluetoothPairingUi FetchDevice() tests.
This CL fixes flaky test in Bluetooth pairing ui page, the test flakes
because when attempting to pair to a device using an address the
waitForFetchDevice() promise will not be resolved.
waitForFetchDevice() creates a promise that is resolved when
pairDevice() is called, but sometimes pairDevice() gets called before
waitForFetchDevice(), so the promise is never resolved. This CL fixes
this by only calling waitForFetchDevice() if completeFetchDevice()
has been called before fetchDevice().

Fixed: b/213943745
Change-Id: Ia16277e594f78b1204822f9ef914029ca022608e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469823
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Commit-Queue: Gordon Seto <gordonseto@google.com>
Cr-Commit-Position: refs/heads/main@{#972201}