multi-instance: Open a link in other window correctly
Opening a link in other window through the context menu had a bug not
setting up the Intent with the flags required to do that. This CL
rectifies it:

- Set activity flags for launching a new instance
- Find the right window to open link in if present

Now not only links but also icon tiles on NTP supports the context
menu option fine.

Bug: 1296484
Change-Id: I6951196b452b180d50f25298ead9428c3def2c42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3468877
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972221}