Add CloudReportingUploadFrequency policy
The policy is used to control the upload interval of CBCM reporting. The
minimum value can be set is 3 hours. Any value that is lower than that
will be bump to it. The default settings remains in 24 hours.

Note that this policy is now also applied to Profile preference so that
it can be reused by Profile reporting. However, whether we need to reuse
the same policy/pref for profile reporting is undecided.

When policy value is updated, any existing pending report will be
rescheduled. If the new frequency is lower than the current waiting
time, a new report will be uploaded immediately.

Change-Id: Id33750392ca8a9d8ab1bea4f7628791330d0bf4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3461872
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972225}