[CrOS Cellular] Add reset euicc functionality in network debug page
This CL adds a reset euicc button in chrome://network general tab. Once
the button is clicked, it will reset the current euicc. If the admin
enables the restrict cellular network or a managed eSIM profile already
installed on the euicc, the reset euicc is not allowed. This
functionality will mainly be used by the certification testing.

Bug: b/210744942
Change-Id: If8376eb59a3f89bf1bb1aaeda6dc9b82b16636bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3453308
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Jason Zhang <jiajunz@google.com>
Cr-Commit-Position: refs/heads/main@{#972231}