Fix productivity launcher kList CTR metrics
We assume every search result view updates the AppListNotifier one
time when they display new results, with all the results it has
just displayed.

But the productivity launcher search result view instead updates the
notifier N times, once for each category, with the results in the
category. This breaks per-type kList metrics.

This CL is a 'quick fix' to treat these results differently. It
doesn't correctly handle a result being _removed_ from the list,
but we know that this only happens on result promotion to best match,
which I think is rare enough that we can ignore it.

On the way I've fixed a bug (that only applies to prod. launcher
metrics) where we weren't clearing |results_| on query change.

Fixed: 1265299
Change-Id: I109fbb91055a08fbf1b4033d81ad26ccda9308c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3466859
Reviewed-by: Rachel Wong <wrong@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972267}