[CrOSCellular] Fix non-active eSIM profiles after powerwash.
This CL fixes an issue with non-active eSIM profiles becoming unmanaged
after a powerwash. This issue was caused because after powerwash, Shill
is only aware of active eSIM profiles and has configuration entries only
for this profile. Policy applicator is unable to match ONC configs by
ICCID since it only looks at configuration entries. This causes such
configs to attempt installation but this never succeeds.

Fixed this by adding a check for existing eSIM profiles in
CellularPolicyHandler. If an existing profile is found, then we perform
the shill configuration step in CellularESimInstaller so that the
Shill service is configured correctly as a managed network.

This also adds additional wait steps in CellularPolicyHandler so
that installation waits for EUICC to become available and profile list
to be refreshed.

Bug: b/218874306
Change-Id: Ibc59bb7e73a4844bb46847bce727f63f22b4ea6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3466366
Reviewed-by: Jason Zhang <jiajunz@google.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Azeem Arshad <azeemarshad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972290}