[HPS] Disable snooping detection in invalid physical configurations.
The models we've produced use training data from "typical" laptop usage.
That is, with the device in laptop mode, directly facing the user on
their desk or lap. We want to disable the feature when the sensor would
receive data particularly different from the training data (in which
case we would not expect high-quality model output).

We use the following two heuristics:

  1) The device must not be in "physically-tablet" mode. That means,
     independent of tablet/non-tablet UI, the device must not be
     physically folded over into a tablet shape. If it is, we expect the
     sensor to see the user at atypical angles.

  2) The device's internal screen must not be rotated in software. We
     assume that the user's orientation always matches the internal
     screen's, and hence the sensor will see a rotated user when the
     internal screen is rotated.

Change-Id: Idfdb86ce187c773828c7b5785119441fff64c5c6
Bug: 1295108
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3459931
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Michael Martis <martis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972329}