[omnibox] Compute a stripped URL if it is missing in OmniboxResult.
The Omnibox provider sometimes receives AutocompleteMatches with empty
stripped_destination_url. These are set as the result ID, but empty IDs
can cause crashes in the launcher.

This CL calculates the stripped URL if it is missing, which should
ensure that the ID is non-empty and still useful for the deduplication
step.

Bug: 1282338
Change-Id: I7eb93387ef48ec0119afd731e1d3d98f9499d8f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3470345
Reviewed-by: Tony Yeoman <tby@chromium.org>
Commit-Queue: Rachel Wong <wrong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972350}