[cropToPromise][#1] Add crop_version in Blink and send to MSDH
BrowserCaptureMediaStreamTrack (BCMST) produces Promises in response
to calls to cropTo(). These Promises can be settled:
1. Immediately, if a patently invalid crop-ID is provided.
2. Shortly after an IPC round-trip to the browser process, if the
   browser process reports that it rejects cropTo().
3. After Viz is informed of the new crop-target, and acknowledges
   that it has received the message and will crop all future
   frames to the right target.

To support #3, we need to be able to resolve these Promises later than
we have been doing so far. For that purpose, this CL stores these
Promises inside the issuing BCMST. Each Promise is associated with
a crop-version. Later CLs will associate frames with the crop-version
they correspond to, which allows Blink to resolve all pending Promises
for a given track once it observes a frame with a newer crop-version.

Bug: 1266378
Change-Id: Ib29eb7d6e5a8009d8e5c6c228ab6e6d12707dd58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469196
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972353}