Don't suppress AX events on live regions which just became unignored
BrowserAccessibilityComWin::UpdateStep3FireEvents() returns early
on ignored nodes as well as on nodes which used to be ignored but
no longer are. Given a live region which was ignored but is now
active, returning early results in no announcement being made.

This change only returns early if the newly-unignored node is not
inside or the root of an active live region.

AX-Relnotes: Fixes bug causing some ARIA alerts to not be presented by
screen readers on Windows.

Bug: 1249045
Change-Id: If4b4f95d9cd2b10747ebe1155992732af09aff7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3452066
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Cr-Commit-Position: refs/heads/main@{#972360}