exo: Restrict pointer lock per-client. Add Overview prompt to Crostini.
* ARC++ continues to permit pointer lock regardless of flags.
* Other clients must have kUseOverviewToExitPointerLock==true set as a
  window property, as well as chrome://flags#exo-pointer-lock enabled.
* Enable kUseOverviewToExitPointerLock for Crostini (and Bruschetta, as
  it doesn't identify its windows separately).

The "Press Overview" prompts are therefore displayed in Crostini and
Bruschetta in addition to Borealis (when the Chrome flag is enabled).

Further, we ensure that no Exo client is able to lock the pointer unless
the "Press Overview" prompt is displayed. The only exceptions to this
rule are ARC++ (through long-standing precedent) and Lacros (which uses
a different mechanism to lock the pointer).

See http://go/exo-pointer-lock-2022 for the rationale and next steps.

BUG=b:203622329

Change-Id: I77cd78b632ad53f6ec5607bff2bfbd2045f89138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469745
Auto-Submit: Chloe Pelling <cpelling@google.com>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972362}