[Autofill Assistant] Add JS snippet action
Adds a new action that takes a controlled JS snippet and executes it
in the context of the webpage. The action is always executed on an
element that can be referenced as `this` in the snippet.

The action may time out if a Promise is not resolved in time.

See launch bug: crbug.com/1293758

Bug: b/203489332
Change-Id: Ibd1957912d122cbd5fd352a9e8b5c336bb9db98f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3046192
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/main@{#972372}