Ensure AwContentsIoThreadClient has valid java_object_
Avoid creating instances of AwContentsIoThreadClient if there is no
valid java delegate to wrap. Callers already expect that the lookup
functions may return nullptr and have to specify default behavior if the
threadclient is not available.
This cuts down on the places default behavior has to be specified in the
code base.

Concerningly, the default for GetSafeBrowsingEnabled was |false| in
AwContentsIoThreadClient but if no client is available,
|AwUrlCheckerDelegateImpl::ShouldSkipRequestCheck| will instead use
|true| as the default value.

Bug: 979321
Change-Id: I8dbaa3891a95f95a68d40745d8afa00c3e7d4651
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3455794
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Peter Pakkenberg <pbirk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972402}