Move byteBuffer.rewind() to background thread
I've been investigating crbug.com/1297894. It looks
like there is a race condition accessing the
WebContentsState ByteBuffer. When CriticalPersistedTabData
is turned on we save both TabState and CriticalPersistedTabData.
byteBuffer.rewind() on getSerializeSupplier in
CriticalPersistedTabData runs on the UI thread. It should run
on the background thread via the SequencedTaskRunner like it
does in TabStateFileManager. Moving to getSerializeSupplier.get()
will achieve this. I haven't proved this is the source of the
race condition but it certainly can cause one.

Bug: 1297894
Change-Id: Id16c3b7ed24f3dae7e54021168826108dc18fbb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469826
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972423}