Fenced Frames: CSP check should only match https: and https://*:*
For opaque fenced frames the URL is opaque to the embedding context,
therefore the embedding context should not tell whether the actual URL
it maps to is allowed or not and should only match the scheme.

Note that opaque fenced frames only allow https urls, therefore we will
only match 'https:' and 'https://*:*'rules. If we do want to support
non-https potentially trustworthy urls, then we might need to extend the
https scheme to match all potentially trustworthy urls to avoid privacy
leak.

This CL introduced a fenced frame mode in FrameTreeNode to indicate
whether the fenced frame url is opaque(urn:uuid) or transparent. The
mode is implicitly inferred from the url of the first navigation and
will stay the same for the lifetime of the fenced frame. A follow-up CL
will enforce that the embedder will not be able to navigate a fenced
frame in an opaque mode to a transparent URL, or vice versa. See the
following docs for the 'mode' attribute to be introduced.

https://docs.google.com/document/d/1HAU9IiHZU4KBPC_rEk3BQYrWTK50PdNGg8CcHhVLXig/edit#heading=h.d4ku3xszbtiv

https://docs.google.com/document/d/1XKBrAINpA6raSR0mMf7_T1RMUyq4JOmNufhUPrs69vA/edit#heading=h.sd3c9yhx0k8

Bug: 1278482
Change-Id: I968f49322d82b56c3f67a79caf3eccf5e4e82fe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3399204
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972480}