Fix InterestGroupFencedFrameBrowserTest.AdComponentsNotLeaked
It runs a FLEDGE auction, with component ads, loads the main ad in a
frame, the component ad in a frame, and then makes sure that the
component ad itself has component ads that all map to about:blank, so
that whether or not the ad is a component ad isn't leaked to that ad,
to minimize information leakage.

Then tries to load the component ad's bogus about:blank component ads
in the component ad's own nested iframe. This step was done
incorrectly - it was loading the component ad's URN URL in the top
frame instead, which wasn't intended, and which only works in the
ShadowDOM case, since in the MPArch case, fenced frames each have
their own URN mapping, while it's per-tab in the ShadowDOM case.

This CL fixes that.

Bug: 1268238
Change-Id: I22461496d217dfd17727fc838acf1705c804c906
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3468942
Auto-Submit: Matt Menke <mmenke@chromium.org>
Reviewed-by: Caleb Raitto <caraitto@chromium.org>
Commit-Queue: Caleb Raitto <caraitto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972484}