End TabDrag sessions before adding a new Tab to the TabStripModel
For multiple cases of TabGroupHeader dragging, when the tab strip model
is changed to have a new tab in the group, the context doesnt currently
pick it up. Because of this, the drag session needs to be ended before
the tab can enter into the TabStripModel. This is because it picks the
index for insertion based on where the group currently is, which may not
be the intended destination for the tab when the drag is in a certain
state.

In order to allow for this, a new signal is created for Observers of the
TabStripModel OnTabWillBeAdded which is called before any mutation
occurs. This allows the drag code to cancel or complete its drag session
before the new tab decides where it wants to go.



Bug: 1280205
Change-Id: Id1e7d752ea1253695bf2f7a229e637415f8523c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3404696
Reviewed-by: Connie Wan <connily@chromium.org>
Commit-Queue: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972499}